% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_dae.R
\name{dae0}
\alias{dae0}
\title{Modules for denoising autoencoders (without latent dim)}
\usage{
dae0(
  n.features,
  input.dropout,
  hidden.dropout,
  encoder.structure,
  decoder.structure,
  act
)
}
\description{
Set up an denoising autoencoder imputer object with specified hyperparameters and then obtain multiple imputed datasets
}
