% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dae1.R
\name{dae1}
\alias{dae1}
\title{Modules for denoising autoencoders}
\usage{
dae1(
  n.features,
  input.dropout,
  latent.dropout,
  hidden.dropout,
  encoder.structure,
  latent.dim,
  decoder.structure,
  act
)
}
\description{
Set up an denoising autoencoder imputer object with specified hyperparameters and then obtain multiple imputed datasets
}
