% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{feature_type}
\alias{feature_type}
\title{This function is used to return the type(numeric,binary,multiclass) of each feature}
\usage{
feature_type(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
This function is used to return the type(numeric,binary,multiclass) of each feature
}
