% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miae-package.R
\docType{package}
\name{miae-package}
\alias{miae}
\alias{miae-package}
\title{\pkg{miae}: Multiple Imputation Through Autoencoders}
\description{
Miae offers a scalable solution for imputing large datasets using various types of autoencoders.
}
\author{
\strong{Maintainer}: Yongshi Deng \email{yongshi.deng@auckland.ac.nz} (\href{https://orcid.org/0000-0001-5845-859X}{ORCID})

Other contributors:
\itemize{
  \item Thomas Lumley \email{t.lumley@auckland.ac.nz} [thesis advisor]
}

}
