% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midae.R
\name{midae}
\alias{midae}
\title{multiple imputation through denoising autoencoders}
\usage{
midae(
  data,
  m = 5,
  epochs = 10,
  dropout.prob = 0.5,
  latent.dim = 16,
  learning.rate = 0.001,
  batch.size = 50,
  encoder.structure = c(128, 64, 32),
  decoder.structure = c(32, 64, 128)
)
}
\arguments{
\item{data}{data with missing values}

\item{m}{the number of imputed datasets}

\item{epochs}{the number of training epochs}

\item{latent.dim}{the size of latent layer}

\item{learning.rate}{learning rate}

\item{batch.size}{the size of each batch}

\item{encoder.structure}{the size of each layer in the encoder}

\item{decoder.structure}{the size of each layer in the decoder}
}
\description{
multiple imputation through denoising autoencoders
}
