% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mivaelatent.R
\name{mivae.latent}
\alias{mivae.latent}
\title{multiple imputation through variational autoencoders (latent) haven't change default setting yet}
\usage{
mivae.latent(
  data,
  m = 5,
  categorical.encoding = "embeddings",
  device = "cpu",
  epochs = 5,
  batch.size = 32,
  subsample = 1,
  early.stopping.epochs = 1,
  vae.params = list(),
  pmm.params = list(),
  loss.na.scale = FALSE,
  verbose = TRUE,
  print.every.n = 1,
  save.model = FALSE,
  path = NULL
)
}
\arguments{
\item{data}{A data frame, tibble or data table with missing values.}

\item{m}{The number of imputed datasets.}

\item{categorical.encoding}{The method for representing multi-class categorical features. Can be either "embeddings" or "onehot" (default).}

\item{device}{Device to use. Either "cpu" (default) or "cuda" for GPU.}

\item{epochs}{The number of training epochs (iterations). Default: 100.}

\item{batch.size}{The size of samples in each batch. Default: 512.}

\item{subsample}{The subsample ratio of training data. Default: 1.}

\item{early.stopping.epochs}{An integer value \code{k}. Mivae training will stop if the validation performance has not improved for \code{k} epochs, only used when \code{subsample}<1. Default: 1.}

\item{vae.params}{A list of parameters for variational autoencoders. See \code{\link[=vae_default]{vae_default}} for details.}

\item{pmm.params}{A list of parameters for predictive mean matching. See \code{\link[=vae_pmm_default]{vae_pmm_default}} for details.}

\item{loss.na.scale}{Whether to multiply the ratio of missing values in  a feature to calculate the loss function. Default: FALSE.}

\item{verbose}{Whether or not to print training loss information. Default: TRUE.}

\item{print.every.n}{If verbose is set to TRUE, print out training loss for every n epochs. Default: 1.}

\item{save.model}{Whether or not to save the imputation model. Default: FALSE.}

\item{path}{The path where the final imputation model will be saved.}
}
\description{
multiple imputation through variational autoencoders (latent) haven't change default setting yet
}
