% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dropout.R
\name{plot_dropout}
\alias{plot_dropout}
\title{plot the density of the impiuted values of a numeric variable using different dropout probabilities}
\usage{
plot_dropout(tune.results, var.name, xlim = NULL, ylim = NULL)
}
\arguments{
\item{tune.results}{object returned by tune_dropout()}

\item{var.name}{the name of a numeric variable}

\item{xlim}{the left and right limit of the x-axis. Default: NULL.}

\item{ylim}{the lower and upper limit of the y-axis. Default: NULL.}
}
\description{
plot the density of the impiuted values of a numeric variable using different dropout probabilities
}
