% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midae.R
\name{pmm_default}
\alias{pmm_default}
\title{Auxiliary function for pmm.params}
\usage{
pmm_default(
  pmm.type = "auto",
  pmm.k = 5,
  pmm.link = "prob",
  pmm.save.vars = NULL
)
}
\description{
Auxiliary function for setting up the default pmm-related parameters for midae and mivae
}
