% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess}
\alias{postprocess}
\title{postprocess data}
\usage{
postprocess(output.data, pre.obj, scaler)
}
\arguments{
\item{output.data}{raw output data from autoencoders}

\item{pre.obj}{preprocess data object}

\item{scaler}{scaler used to scale the data}
}
\description{
postprocess data
}
