% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Pre-process data before imputation: scale numeric and one-hot categorical}
\usage{
preprocess(
  data,
  scaler = "none",
  lower,
  upper,
  categorical.encoding = "embeddings",
  initial.imp = "sample"
)
}
\description{
Pre-process data before imputation: scale numeric and one-hot categorical
}
