% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scaler.R
\name{rev_decile_scaler}
\alias{rev_decile_scaler}
\title{This function back-transform decile-scaled data to an output as data matrix}
\usage{
rev_decile_scaler(scaled.data, num.names, decile1, decile9)
}
\arguments{
\item{scaled.data}{A matrix or array with scaled numeric data}

\item{num.names}{the names of numeric features}

\item{decile1}{A vector that contains the the 1st decile (10th percentile) of each column}

\item{decile9}{A vector that contains the the 9th decile (90th percentile) of each column}
}
\description{
This function back-transform decile-scaled data to an output as data matrix
}
