% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scaler.R
\name{rev_minmax_scaler}
\alias{rev_minmax_scaler}
\title{This function back-transform minmax-scaled data to an output as data matrix}
\usage{
rev_minmax_scaler(scaled.data, num.names, colmin, colmax)
}
\arguments{
\item{scaled.data}{A matrix or array with scaled numeric data}

\item{num.names}{the names of numeric features}

\item{colmin}{A vector that contains the minimum of each column}

\item{colmax}{A vector that contains the maximum of each column}
}
\description{
This function back-transform minmax-scaled data to an output as data matrix
}
