% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{rev_onehot}
\alias{rev_onehot}
\title{reverse onehot data}
\usage{
rev_onehot(onehot.data, pre.obj)
}
\arguments{
\item{onehot.data}{data with onehot features}

\item{pre.obj}{preprocess data object}
}
\description{
reverse onehot data
}
