% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scaler.R
\name{rev_robust_scaler}
\alias{rev_robust_scaler}
\title{This function back-transform robust-scaled data to an output as data matrix}
\usage{
rev_robust_scaler(
  scaled.data,
  num.names,
  robust.lower,
  robust.upper,
  robust.median
)
}
\arguments{
\item{scaled.data}{A matrix or array with scaled numeric data}

\item{num.names}{the names of numeric features}

\item{robust.lower}{A vector that contains the the lower quantile (25\% by default) of each column}

\item{robust.upper}{A vector that contains the the upper quantile (75\% by default) of each column}

\item{robust.median}{A vector that contains the the median (50\%) of each column}
}
\description{
This function back-transform robust-scaled data to an output as data matrix
}
