% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scaler.R
\name{rev_standard_scaler}
\alias{rev_standard_scaler}
\title{This function back-transform standardized data to an output as data matrix}
\usage{
rev_standard_scaler(scaled.data, num.names, colmean, colsd)
}
\arguments{
\item{scaled.data}{A matrix or array with scaled numeric data}

\item{num.names}{the names of numeric features}

\item{colmean}{A vector that contains the mean of each column}

\item{colsd}{A vector that contains the standard deviation of each column}
}
\description{
This function back-transform standardized data to an output as data matrix
}
