% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scaler.R
\name{robust}
\alias{robust}
\title{scale a vector using robust scaler}
\usage{
robust(x, lower = 0.25, upper = 0.75)
}
\arguments{
\item{x}{a vector}

\item{lower}{The lower quantile (25\% by default)}

\item{upper}{The upper quantile (75\% by default)}
}
\description{
scale a vector using robust scaler
}
