% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scaler.R
\name{standard_scaler}
\alias{standard_scaler}
\title{scale a dataset using stardardize and return a scaled dataframe, the colmean and colsd of each column}
\usage{
standard_scaler(data)
}
\arguments{
\item{data}{A data frame or tibble}
}
\description{
scale a dataset using stardardize and return a scaled dataframe, the colmean and colsd of each column
}
