% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_vae_separate.R
\name{vae.encoder}
\alias{vae.encoder}
\title{Modules for variational autoencoders (separate encoder and decoder)}
\usage{
vae.encoder(
  categorical.encoding,
  n.others,
  cardinalities,
  embedding.dim,
  input.dropout,
  hidden.dropout,
  encoder.structure,
  latent.dim,
  act
)
}
\description{
variational autoencoders (separate encoder and decoder)
}
