% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_vae.R
\name{vae}
\alias{vae}
\title{Modules for variational autoencoders}
\usage{
vae(
  categorical.encoding,
  n.others,
  cardinalities,
  embedding.dim,
  input.dropout,
  hidden.dropout,
  encoder.structure,
  latent.dim,
  decoder.structure,
  act
)
}
\description{
Set up a autoencoder imputer object with specified hyperparameters and then obtain multiple imputed datasets
}
