% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mivae.R, R/mivaelatent.R
\name{vae_default}
\alias{vae_default}
\title{Auxiliary function for vae.params}
\usage{
vae_default(
  shuffle = TRUE,
  drop.last = FALSE,
  beta = 1,
  input.dropout = 0,
  hidden.dropout = 0,
  optimizer = "adamW",
  learning.rate = 0.001,
  weight.decay = 0.01,
  momentum = 0,
  dampening = 0,
  eps = 1e-08,
  rho = 0.9,
  alpha = 0.99,
  learning.rate.decay = 0,
  encoder.structure = c(256, 128, 64),
  latent.dim = 8,
  decoder.structure = c(64, 128, 256),
  act = "relu",
  init.weight = "he.normal",
  scaler = "standard",
  initial.imp = "sample",
  lower = 0.25,
  upper = 0.75
)

vae_default(
  shuffle = TRUE,
  drop.last = FALSE,
  beta = 1,
  input.dropout = 0,
  hidden.dropout = 0,
  optimizer = "adamW",
  learning.rate = 0.001,
  weight.decay = 0.01,
  momentum = 0,
  dampening = 0,
  eps = 1e-08,
  rho = 0.9,
  alpha = 0.99,
  learning.rate.decay = 0,
  encoder.structure = c(256, 128, 64),
  latent.dim = 8,
  decoder.structure = c(64, 128, 256),
  act = "relu",
  init.weight = "he.normal",
  scaler = "standard",
  initial.imp = "sample",
  lower = 0.25,
  upper = 0.75
)
}
\description{
Auxiliary function for setting up the default vae-related hyperparameters for mivae

Auxiliary function for setting up the default vae-related hyperparameters for mivae
}
