% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mivae.R
\name{vae_pmm_default}
\alias{vae_pmm_default}
\title{Auxiliary function for pmm.params for mivae}
\usage{
vae_pmm_default(
  pmm.type = NULL,
  pmm.k = 5,
  pmm.link = "prob",
  pmm.save.vars = NULL
)
}
\description{
Auxiliary function for setting up the default pmm-related parameters for mivae
}
