% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_dae.R
\name{dae}
\alias{dae}
\title{Modules for denoising autoencoders}
\usage{
dae(n.features, latent.dim, dropout.prob, encoder.structure, decoder.structure)
}
\description{
Set up an denoising autoencoder imputer object with specified hyperparameters and then obtain multiple imputed datasets
}
