% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_dae.R
\name{dae}
\alias{dae}
\title{Modules for denoising autoencoders}
\usage{
dae(
  categorical.encoding,
  n.others,
  cardinalities,
  embedding.dim,
  input.dropout,
  hidden.dropout,
  encoder.structure,
  latent.dim,
  decoder.structure,
  act
)
}
\description{
Set up an denoising autoencoder imputer object with specified hyperparameters and then obtain multiple imputed datasets
}
