% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midae.R
\name{dae_default}
\alias{dae_default}
\title{Auxiliary function for dae.params}
\usage{
dae_default(
  shuffle = TRUE,
  drop.last = FALSE,
  input.dropout = 0.2,
  hidden.dropout = 0.5,
  optimizer = "adamW",
  learning.rate = 0.001,
  weight.decay = 0.01,
  momentum = 0,
  dampening = 0,
  eps = 1e-08,
  rho = 0.9,
  alpha = 0.99,
  learning.rate.decay = 0,
  encoder.structure = c(128, 64, 32),
  latent.dim = 16,
  decoder.structure = c(32, 64, 128),
  act = "relu",
  init.weight = "he.normal",
  scaler = "standard",
  initial.imp = "sample",
  lower = 0.25,
  upper = 0.75
)
}
\description{
Auxiliary function for setting up the default dae-related hyperparameters for midae
}
