% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midae.R
\name{dae_pmm_default}
\alias{dae_pmm_default}
\title{Auxiliary function for pmm.params for midae}
\usage{
dae_pmm_default(
  pmm.type = NULL,
  pmm.k = 5,
  pmm.link = "prob",
  pmm.save.vars = NULL
)
}
\arguments{
\item{pmm.type}{The type of predictive mean matching (PMM). Possible values:
\itemize{
\item \code{NULL}: Imputations without PMM;
\item \code{0}: Imputations with PMM type 0;
\item \code{1}: Imputations with PMM type 1;
\item \code{2}: Imputations with PMM type 2;
\item \code{"auto"} (default): Imputations with PMM type 2 for numeric/integer variables; imputations without PMM for categorical variables.
}}

\item{pmm.k}{The number of donors for predictive mean matching. Default: 5}

\item{pmm.link}{The link for predictive mean matching in binary variables
\itemize{
\item \code{"prob"} (default): use probabilities;
\item \code{"logit"}: use logit values.
}}

\item{pmm.save.vars}{The names of variables whose predicted values of observed entries will be saved. Only use for PMM.}
}
\description{
Auxiliary function for setting up the default pmm-related parameters for midae
}
