% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scaler.R
\name{decile_scaler}
\alias{decile_scaler}
\title{scale a dataset using decile and return a scaled dataframe, the 1st decile and the 9th decile of each column}
\usage{
decile_scaler(data)
}
\arguments{
\item{data}{A data frame or tibble}
}
\description{
scale a dataset using decile and return a scaled dataframe, the 1st decile and the 9th decile of each column
}
