% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_new.R
\name{impute_new}
\alias{impute_new}
\title{Impute new data with a saved \code{midae} or \code{mivae} imputation model}
\usage{
impute_new(object, newdata, pmm.k = NULL, m = NULL, verbose = FALSE)
}
\arguments{
\item{object}{A saved imputer object created by \code{midae(..., save.models = TRUE)} or \code{mivae(..., save.models = TRUE)}.}

\item{newdata}{A data frame, tibble or data.table. New data with missing values.}

\item{pmm.k}{The number of donors for predictive mean matching. If \code{NULL} (the default), the \code{pmm.k} value in the saved imputer object will be used.}

\item{m}{The number of imputed datasets. If \code{NULL} (the default), the \code{m} value in the saved imputer object will be used.}

\item{verbose}{A logical value indicating whether to print the progress.}
}
\value{
A list of \code{m} imputed datasets for new data.
}
\description{
Impute new data with a saved \code{midae} or \code{mivae} imputation model
}
