% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scaler.R
\name{minmax_scaler}
\alias{minmax_scaler}
\title{scale a dataset using minmax and return a scaled dataframe, the colmin and colmax of each column}
\usage{
minmax_scaler(data, initial.imp = "sample")
}
\arguments{
\item{data}{A data frame or tibble}
}
\description{
scale a dataset using minmax and return a scaled dataframe, the colmin and colmax of each column
}
