% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_scaler.R
\name{robust_scaler}
\alias{robust_scaler}
\title{scale a dataset using robust scaler and return a scaled dataframe, the median, the lower quantile (25\% by default) and the upper quantile (75\% by default) of each column}
\usage{
robust_scaler(data, initial.imp = "sample", lower = 0.25, upper = 0.75)
}
\arguments{
\item{data}{A data frame or tibble}

\item{initial.imp}{The method for initial imputation. Can be "mean", "median" or "sample" (default).}

\item{lower}{The lower quantile (25\% by default)}

\item{upper}{The upper quantile (75\% by default)}
}
\description{
scale a dataset using robust scaler and return a scaled dataframe, the median, the lower quantile (25\% by default) and the upper quantile (75\% by default) of each column
}
