% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_vae_dropout.R
\name{tune_vae_dropout}
\alias{tune_vae_dropout}
\title{tuning dropout for mivae}
\usage{
tune_vae_dropout(
  data,
  m = 5,
  epochs = 10,
  batch.size = 50,
  dropout.grid = dropout.grid,
  latent.dropout = 0,
  optim = "adam",
  learning.rate = 0.001,
  weight.decay = 0,
  momentum = 0,
  encoder.structure = c(128, 64, 32),
  latent.dim = 8,
  decoder.structure = c(32, 64, 128),
  verbose = TRUE,
  print.every.n = 1,
  save.model = FALSE,
  path = NULL
)
}
\arguments{
\item{data}{A data frame, tibble or data table with missing values.}

\item{m}{The number of imputed datasets.}

\item{epochs}{The number of training epochs (iterations).}

\item{batch.size}{The size of samples in each batch.}

\item{dropout.grid}{A list of tuning values for input.dropout and hidden.dropout}

\item{latent.dropout}{The dropout probability of the latent layer.}

\item{optim}{The name of the optimizer. Options are : "adam" (default), and "sgd".}

\item{learning.rate}{The learning rate. The default value is 0.001.}

\item{weight.decay}{Weight decay (L2 penalty). The default value is 0.}

\item{momentum}{Parameter for "sgd" optimizer. It is used for accelerating SGD in the relevant direction and dampens oscillations.}

\item{encoder.structure}{A vector indicating the structure of encoder. Default: c(128,64,32)}

\item{latent.dim}{The size of latent layer. The default value is 8.}

\item{decoder.structure}{A vector indicating the structure of decoder. Default: c(32,64,128)}

\item{verbose}{Whether or not to print training loss information. Default: TRUE.}

\item{print.every.n}{If verbose is set to TRUE, print out training loss for every n epochs.}

\item{save.model}{Whether or not to save the imputation models for different combinations of dropouts. Default: FALSE.}

\item{path}{The path where the final imputation model will be saved.}
}
\description{
tuning dropout for mivae
}
