% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mivae.R
\name{vae_default}
\alias{vae_default}
\title{Auxiliary function for vae.params}
\usage{
vae_default(
  shuffle = TRUE,
  drop.last = FALSE,
  beta = 1,
  input.dropout = 0,
  hidden.dropout = 0,
  optimizer = "adamW",
  learning.rate = 1e-04,
  weight.decay = 0.002,
  momentum = 0,
  eps = 1e-07,
  encoder.structure = c(128, 64, 32),
  latent.dim = 16,
  decoder.structure = c(32, 64, 128),
  act = "elu",
  init.weight = "xavier.normal",
  scaler = "standard",
  initial.imp = "sample",
  lower = 0.25,
  upper = 0.75
)
}
\description{
Auxiliary function for setting up the default vae-related hyperparameters for mivae
}
